fix(openrpc): set subpackages in openrpc
