+++ packages/parsers/src/openapi/3.1/OpenApiDocumentConverter.node.ts
@@ -1,7 +1,7 @@
-import { camelCase } from "es-toolkit";
 import { OpenAPIV3_1 } from "openapi-types";
 import { v4 } from "uuid";
 import { FernRegistry } from "../../client/generated";
+import { computeSubpackages } from "../../utils/computeSubpackages";
 import {
   BaseOpenApiV3_1ConverterNode,
   BaseOpenApiV3_1ConverterNodeConstructorArgs,
@@ -146,35 +146,7 @@ export class OpenApiDocumentConverterNode extends BaseOpenApiV3_1ConverterNode<
     const subpackages: Record<
       FernRegistry.api.v1.SubpackageId,
       FernRegistry.api.latest.SubpackageMetadata
-    > = {};
-    if (endpoints != null) {
-      Object.values(endpoints).forEach((endpoint) =>
-        endpoint.namespace?.forEach((subpackage) => {
-          const qualifiedPath: string[] = [];
-          subpackages[FernRegistry.api.v1.SubpackageId(camelCase(subpackage))] =
-            {
-              id: FernRegistry.api.v1.SubpackageId(camelCase(subpackage)),
-              name: [...qualifiedPath, subpackage].join("/"),
-              displayName: subpackage,
-            };
-          qualifiedPath.push(subpackage);
-        })
-      );
-    }
-    if (webhookEndpoints != null) {
-      Object.values(webhookEndpoints).forEach((webhook) =>
-        webhook.namespace?.forEach((subpackage) => {
-          const qualifiedPath: string[] = [];
-          subpackages[FernRegistry.api.v1.SubpackageId(camelCase(subpackage))] =
-            {
-              id: FernRegistry.api.v1.SubpackageId(camelCase(subpackage)),
-              name: [...qualifiedPath, subpackage].join("/"),
-              displayName: subpackage,
-            };
-          qualifiedPath.push(subpackage);
-        })
-      );
-    }
+    > = computeSubpackages({ endpoints, webhookEndpoints });
 
     const types = this.components?.convert();
